+++ src/libcore/MemObj.cpp
@@ -67,7 +67,7 @@ void MemObj::computenUpperCaches()
 {
   nUpperCaches = 0;
 
-  for(ulong j=0; j < upperLevel.size(); j++) {
+  for(uint j=0; j < upperLevel.size(); j++) {
     if(upperLevel[j]->isCache())
       nUpperCaches++;
     else 
@@ -75,7 +75,7 @@ void MemObj::computenUpperCaches()
   }
 
   // top-down traversal of the memory objects
-  for(ulong i=0; i < lowerLevel.size(); i++) {
+  for(uint i=0; i < lowerLevel.size(); i++) {
     lowerLevel[i]->computenUpperCaches();
   }
 }
